feat: refactor drawer in components dir
